Рефакторинг: разделение сторов по компонентам
