+++ client/src/redux/authToolkitRedux/StoreSlices/index.ts
@@ -1,14 +1,18 @@
 import { combineReducers } from 'redux'
 
 import { store } from 'components/App/AppBehavior'
-import { app } from './app'
-import { authorization } from './authorization'
 import { registration } from './registration'
+import { user } from 'pages/Profile/userSlice'
+import { feed } from 'components/Feed/feedSlice'
+import { upload } from 'pages/AddMeme/uploadSlice'
+import { authorization } from 'pages/Authorization/authSlice'
 
 export const reducers = combineReducers({
   authorization,
   registration,
-  app,
+  user,
+  feed,
+  upload,
 })
 
 // export type RootState = ReturnType<typeof reducers>
